+++ pkg/tenantcluster/factory.go
@@ -22,6 +22,13 @@ type tenantClientFactory struct {
 }
 
 func NewFactory(certsSearcher certs.Interface, logger micrologger.Logger) (Factory, error) {
+	if certsSearcher == nil {
+		return nil, microerror.Maskf(invalidConfigError, "certsSearcher must not be empty")
+	}
+	if logger == nil {
+		return nil, microerror.Maskf(invalidConfigError, "logger must not be empty")
+	}
+
 	c := tenantcluster.Config{
 		CertsSearcher: certsSearcher,
 		Logger:        logger,
@@ -43,11 +50,12 @@ func NewFactory(certsSearcher certs.Interface, logger micrologger.Logger) (Facto
 }
 
 func (tcf *tenantClientFactory) GetClient(ctx context.Context, cr *capiv1alpha3.Cluster) (client.Client, error) {
+	tcf.logger.LogCtx(ctx, "level", "debug", "message", "Creating tenant cluster k8s client for cluster %#q", key.ClusterID(cr))
 	var k8sClient k8sclient.Interface
 	{
 		restConfig, err := tcf.tenantRestConfigProvider.NewRestConfig(ctx, key.ClusterID(cr), cr.Spec.ControlPlaneEndpoint.String())
 		if tenant.IsAPINotAvailable(err) {
-			return nil, microerror.Mask(APINotAvailableError)
+			return nil, microerror.Mask(apiNotAvailableError)
 		} else if err != nil {
 			return nil, microerror.Mask(err)
 		}